refactor: container stats command
