+++ pkg/cmd/container/stats_freebsd.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package container
 
 import (
 	"github.com/containerd/nerdctl/pkg/inspecttypes/native"
