+++ pkg/cmd/container/stats_linux.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package container
 
 import (
 	"errors"
